pkg/cover: optimize inline function logic
