+++ pkg/cover/backend/elf.go
@@ -4,12 +4,10 @@
 package backend
 
 import (
-	"debug/dwarf"
 	"debug/elf"
 	"encoding/binary"
 	"fmt"
 	"io"
-	"sort"
 	"strings"
 
 	"github.com/google/syzkaller/pkg/host"
@@ -25,215 +23,6 @@ func makeELF(target *targets.Target, objDir, srcDir, buildDir string,
 			readModuleCoverPoints: elfReadModuleCoverPoints,
 		},
 	)
-}
-
-type DwarfCompileUnit struct {
-	Entry  *dwarf.Entry
-	Name   string
-	Ranges [][2]uint64
-	Offset dwarf.Offset
-}
-
-type DwarfFunction struct {
-	DwarfCompileUnit *DwarfCompileUnit
-	Type             dwarf.Tag
-	Name             string
-	Ranges           [][2]uint64
-	DeclFile         string
-	AbstractOrigin   dwarf.Offset
-	Inline           bool
-	Offset           dwarf.Offset
-}
-
-func getFilenameByIndex(debugInfo *dwarf.Data, entry *dwarf.Entry, index int) (string, error) {
-	r, err := debugInfo.LineReader(entry)
-	if err != nil {
-		return "", fmt.Errorf("not found line reader for Compile Unit")
-	}
-	files := r.Files()
-	if files == nil {
-		return "", fmt.Errorf("files == nil")
-	}
-	if index >= len(files) {
-		return "", fmt.Errorf("index (%v) >= len(files) (%v)", index, len(files))
-	}
-	if index == 0 {
-		return "", nil
-	}
-	return files[index].Name, nil
-}
-
-func readAllCompileUnits(debugInfo *dwarf.Data) ([]*DwarfCompileUnit, error) {
-	var data []*DwarfCompileUnit
-
-	for r := debugInfo.Reader(); ; {
-		ent, err := r.Next()
-		if err != nil {
-			return nil, err
-		}
-		if ent == nil {
-			break
-		}
-		if ent.Tag != dwarf.TagCompileUnit {
-			return nil, fmt.Errorf("found unexpected tag %v on top level", ent.Tag)
-		}
-		attrName := ent.Val(dwarf.AttrName)
-		if attrName == nil {
-			continue
-		}
-		ranges, err := debugInfo.Ranges(ent)
-		if err != nil {
-			return nil, err
-		}
-		data = append(data, &DwarfCompileUnit{
-			Entry:  ent,
-			Name:   attrName.(string),
-			Ranges: ranges,
-			Offset: ent.Offset,
-		})
-		r.SkipChildren()
-	}
-
-	return data, nil
-}
-
-func getEntryByOffset(debugInfo *dwarf.Data, offset dwarf.Offset) (*dwarf.Entry, error) {
-	r := debugInfo.Reader()
-	r.Seek(offset)
-	ent, err := r.Next()
-	if err != nil {
-		return nil, err
-	}
-	return ent, nil
-}
-
-func readAllSubprograms(debugInfo *dwarf.Data, compileUnit *DwarfCompileUnit) ([]*DwarfFunction, error) {
-	var data []*DwarfFunction
-
-	top := true
-	first := true
-	for r := debugInfo.Reader(); ; {
-		if top {
-			r.Seek(compileUnit.Offset)
-			top = false
-		}
-		ent, err := r.Next()
-		if err != nil {
-			return nil, err
-		}
-		if ent == nil {
-			break
-		}
-		if first && ent.Tag == dwarf.TagCompileUnit {
-			first = false
-			continue
-		} else if ent.Tag == dwarf.TagCompileUnit {
-			break
-		}
-		if ent.Tag != dwarf.TagSubprogram {
-			continue
-		}
-
-		attrName := ent.Val(dwarf.AttrName)
-		attrAbstractOrigin := ent.Val(dwarf.AttrAbstractOrigin)
-		var decfile string
-		if attrName == nil && attrAbstractOrigin != nil {
-			ent1, err := getEntryByOffset(debugInfo, attrAbstractOrigin.(dwarf.Offset))
-			if err != nil {
-				return nil, err
-			}
-			attrName = ent1.Val(dwarf.AttrName)
-			decfile, err = getFilenameByIndex(debugInfo, compileUnit.Entry, int(ent1.Val(dwarf.AttrDeclFile).(int64)))
-			if err != nil {
-				return nil, err
-			}
-		}
-		if attrName == nil {
-			continue
-		}
-		if decfile == "" && ent.Val(dwarf.AttrDeclFile) != nil {
-			decfile, err = getFilenameByIndex(debugInfo, compileUnit.Entry, int(ent.Val(dwarf.AttrDeclFile).(int64)))
-			if err != nil {
-				return nil, err
-			}
-		}
-		ranges, err := debugInfo.Ranges(ent)
-		if err != nil {
-			return nil, err
-		}
-		inline := false
-		attrInline := ent.Val(dwarf.AttrInline)
-		if attrInline != nil {
-			inline = true
-		}
-
-		data = append(data, &DwarfFunction{
-			DwarfCompileUnit: compileUnit,
-			Type:             dwarf.TagSubprogram,
-			Name:             attrName.(string),
-			Ranges:           ranges,
-			DeclFile:         decfile,
-			Inline:           inline,
-			Offset:           ent.Offset,
-		})
-		r.SkipChildren()
-	}
-
-	return data, nil
-}
-
-func readAllInlinedSubroutines(debugInfo *dwarf.Data, inlineSubprograms map[dwarf.Offset]*DwarfFunction,
-	subprogram *DwarfFunction) ([]*DwarfFunction, error) {
-	var data []*DwarfFunction
-
-	top := true
-	first := true
-	for r := debugInfo.Reader(); ; {
-		if top {
-			r.Seek(subprogram.Offset)
-			top = false
-		}
-		ent, err := r.Next()
-		if err != nil {
-			return nil, err
-		}
-		if ent == nil {
-			break
-		}
-		if first && ent.Tag == dwarf.TagSubprogram {
-			first = false
-			continue
-		} else if ent.Tag == dwarf.TagSubprogram {
-			break
-		}
-		if ent.Tag != dwarf.TagInlinedSubroutine {
-			continue
-		}
-		attrAbstractOrigin := ent.Val(dwarf.AttrAbstractOrigin)
-		if attrAbstractOrigin == nil {
-			continue
-		}
-		df := inlineSubprograms[attrAbstractOrigin.(dwarf.Offset)]
-		ranges, err := debugInfo.Ranges(ent)
-		if err != nil {
-			return nil, err
-		}
-		if len(ranges) == 0 {
-			continue
-		}
-		data = append(data, &DwarfFunction{
-			DwarfCompileUnit: df.DwarfCompileUnit,
-			Type:             dwarf.TagInlinedSubroutine,
-			Name:             df.Name,
-			Ranges:           ranges,
-			Offset:           ent.Offset,
-			DeclFile:         df.DeclFile,
-			Inline:           df.Inline,
-			AbstractOrigin:   attrAbstractOrigin.(dwarf.Offset),
-		})
-	}
-
-	return data, nil
 }
 
 func elfReadSymbols(module *Module, info *symbolInfo) ([]*Symbol, error) {
@@ -255,27 +44,37 @@ func elfReadSymbols(module *Module, info *symbolInfo) ([]*Symbol, error) {
 	}
 	info.textAddr = text.Addr
 	for i, symb := range allSymbols {
-		flag := symb.Value >= text.Addr && symb.Value+symb.Size <= text.Addr+text.Size
+		text := symb.Value >= text.Addr && symb.Value+symb.Size <= text.Addr+text.Size
 		if strings.HasPrefix(symb.Name, "__sanitizer_cov_trace_") {
 			if symb.Name == "__sanitizer_cov_trace_pc" {
 				info.tracePCIdx[i] = true
-				if flag {
+				if text {
 					info.tracePC = symb.Value
 				}
 			} else {
 				info.traceCmpIdx[i] = true
-				if flag {
+				if text {
 					info.traceCmp[symb.Value] = true
 				}
 			}
 		}
 	}
 
-	compileUnits, err := readAllCompileUnits(debugInfo)
-	if err != nil {
-		return nil, err
+	textSecIdx := 0
+	for i, sec := range file.Sections {
+		if sec.Name == ".text" {
+			textSecIdx = i
+		}
 	}
-	allDwarfFunctions, err := elfReadSymbolsFromDwarf(debugInfo, compileUnits)
+	textSymbols := make(map[string]bool)
+	for _, s := range allSymbols {
+		if s.Info&0xf == uint8(elf.STT_FUNC) {
+			if s.Section == elf.SectionIndex(textSecIdx) {
+				textSymbols[s.Name] = true
+			}
+		}
+	}
+	allDwarfFunctions, err := readSymbolsFromDwarf(debugInfo, textSymbols)
 	if err != nil {
 		return nil, err
 	}
@@ -293,17 +92,15 @@ func elfReadSymbols(module *Module, info *symbolInfo) ([]*Symbol, error) {
 				r[1] + module.Addr,
 			})
 		}
-		flag := start >= text.Addr && end <= text.Addr+text.Size
-		if flag {
-			start1 := start + module.Addr
-			end1 := end + module.Addr
+		text := start >= text.Addr && end <= text.Addr+text.Size
+		if text {
 			symbols = append(symbols, &Symbol{
 				Module: module,
 				ObjectUnit: ObjectUnit{
 					Name: ds.Name,
 				},
-				Start:  start1,
-				End:    end1,
+				Start:  start + module.Addr,
+				End:    end + module.Addr,
 				Ranges: ranges,
 				Inline: ds.Inline,
 				Unit: &CompileUnit{
@@ -316,67 +113,6 @@ func elfReadSymbols(module *Module, info *symbolInfo) ([]*Symbol, error) {
 		}
 	}
 	return symbols, nil
-}
-
-func elfReadSymbolsFromDwarf(debugInfo *dwarf.Data, compileUnits []*DwarfCompileUnit) ([]*DwarfFunction, error) {
-	var allSubprograms []*DwarfFunction
-	var allInlineSubroutines []*DwarfFunction
-	for _, cu := range compileUnits {
-		errc := make(chan error, 1)
-		go func(cu *DwarfCompileUnit) {
-			subprograms, err := readAllSubprograms(debugInfo, cu)
-			if err != nil {
-				errc <- err
-				return
-			}
-			allSubprograms = append(allSubprograms, subprograms...)
-			inlineSubprograms := make(map[dwarf.Offset]*DwarfFunction)
-			for _, sp := range subprograms {
-				if sp.Inline {
-					inlineSubprograms[sp.Offset] = sp
-				}
-			}
-			for _, sp := range subprograms {
-				if sp.Inline {
-					continue
-				}
-				erric := make(chan error, 1)
-				go func(sp *DwarfFunction) {
-					inlineSubroutines, err := readAllInlinedSubroutines(debugInfo, inlineSubprograms, sp)
-					if err != nil {
-						erric <- err
-						return
-					}
-					allInlineSubroutines = append(allInlineSubroutines, inlineSubroutines...)
-					erric <- nil
-				}(sp)
-				if err := <-erric; err != nil {
-					errc <- err
-				}
-			}
-			errc <- nil
-		}(cu)
-		if err := <-errc; err != nil {
-			return nil, err
-		}
-	}
-	var allDwarfFunctions []*DwarfFunction
-	for _, df := range allSubprograms {
-		if len(df.Ranges) == 0 {
-			continue
-		}
-		allDwarfFunctions = append(allDwarfFunctions, df)
-	}
-	for _, df := range allInlineSubroutines {
-		if len(df.Ranges) == 0 {
-			continue
-		}
-		allDwarfFunctions = append(allDwarfFunctions, df)
-	}
-	sort.Slice(allDwarfFunctions, func(i, j int) bool {
-		return allDwarfFunctions[i].Ranges[0][0] < allDwarfFunctions[j].Ranges[0][0]
-	})
-	return allDwarfFunctions, nil
 }
 
 func elfReadTextData(module *Module) ([]byte, error) {
